*** empty log message ***
git-svn-id: https://svn.code.sf.net/p/cspfj/code/trunk/cspfj@45 b3bf8218-cbbe-40ea-ae7b-cf512e877a40

+++ src/cspfj/Solver.java
@@ -28,7 +28,7 @@ import cspfj.problem.Variable;
 
 public interface Solver {
 
-    boolean run(final int maxDuration) throws OutOfTimeException, IOException;
+    boolean run(final long maxDuration) throws OutOfTimeException, IOException;
 
     Map<Variable, Integer> getSolution();
